feat(#86): seperate category logic into category directory
